pyproject.toml
[tool.poetry]
name = "backendteste"
version = "0.1.0"
description = ""
authors = ["Jeff <jeffersonpereira138@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
celery = "^5.4.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
